loader: sketch_header: moved sketch header
moved sketch header outside of loader main, in order to favor reuse of
that code section, it can then be used by the sketch itself for ota
